### 2025-01-16 #7675 @editorialbot
[PRE REVIEW]: EWDM: A wavelet-based method for estimating directional spectra of ocean waves
**Submitting author:** @dspelaez (<a href="http://orcid.org/0000-0001-5862-6194">Daniel Peláez-Zapata</a>)
**Repository:** https://github.com/dspelaez/extended-wdm
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
