### 2025-03-28 dspelaez/extended-wdm#2 @rcaneill
[JOSS review] Unused test file
Hi,

Romain (@rcaneill), one of your reviewer for the JOSS submission here.

It seems to me that this file is not used in the tests. Is it an old file that you forgot to remove?

https://github.com/dspelaez/extended-wdm/blob/b27fbeadc1ee68e49d18269f321acf8f78a4f747/tests/test_main.py#L1-L9

https://github.com/openjournals/joss-reviews/issues/7942
