### 2025-03-28 dspelaez/extended-wdm#3 @rcaneill
[JOSS review] missing instruction for development environment
I did not find instruction on how to set up a dev/test environment. Adding `pip install -e .[test]` in the README would be sufficient.

https://github.com/openjournals/joss-reviews/issues/7942
