### 2025-03-28 dspelaez/extended-wdm#3 @rcaneill
Mentioning how to run the tests is also missing, i.e. `pytest tests`
