### 2025-03-28 dspelaez/extended-wdm#4 @rcaneill
[JOSS review] Quick example figure incorrect
The figure produced by the quick example code snippet is not exactly the same as the one in the README: no contour lines are produced for me. Is it a change in the library, or is the figures shown in the README produced with another function/arguments?

![Image](https://github.com/user-attachments/assets/a06614cb-2700-4dd1-9893-4e4ff3ff3637)

https://github.com/openjournals/joss-reviews/issues/7942
