### 2025-03-28 dspelaez/extended-wdm#5 @rcaneill
[JOSS review] Missing community guideline
The item in the review checklist is:
>  Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Adding few lines in the README or documentation would resolve this issue

https://github.com/openjournals/joss-reviews/issues/7942
