### 2025-03-28 #7942 @rcaneill
Hi @rwegener2 @dspelaez, I have been through the checklist once. I have not checked some points in the software part, resolving the issues I opened should be sufficient for me to check these points.

I have not checked the `summary` part of the paper for the following reasons:
* A clear description of the software is given, however it is oriented towards specialists of the field (at least physical oceanographers with a good background)
* Some acronyms, even if known by many oceanographers, are not detailed (e.g. ADCP, CDIP database)
* I feel that it is missing a few sentences / a paragraph at the beginning to introduce the subject to a broad audience
* I understand that it may overlap with the statement of need, but I am confident that without too much rewriting it is possible to better introduce the package

Overall, I found both the package and the paper to be already quite good.
