### 2025-03-28 dspelaez/extended-wdm#6 @rcaneill
[JOSS review] Remove unused Fourier file?
If this file is not used, it should be removed:

https://github.com/dspelaez/extended-wdm/blob/b27fbeadc1ee68e49d18269f321acf8f78a4f747/ewdm/fourier.py#L1-L141

https://github.com/openjournals/joss-reviews/issues/7942
