### 2025-03-31 039c263
unused fourier.py file was deleted, installation of test environment was added to the README.md file
