+++ ewdm/fourier.py
@@ -1,144 +1,0 @@
-#! /usr/bin/env python
-# -*- coding: utf-8 -*-
-# vim:fenc=utf-8
-
-import xarray as xr
-import numpy as np
-from scipy import signal
-
-from .helpers import get_sampling_frequency
-
-
-# NPERSEG = 256
-
-# # {{{
-# class ClassicSpectralAnalysis(object):
-    # """This class implements the classic spectral directional analysis"""
-
-    # def __init__(self, dataset: xr.Dataset, fs: float = 1.0):
-
-        # # try to interpolate before computing fft
-        # if (ds["z"].isnull().sum() / len(ds["time"])) > MAX_NAN_RATIO:
-            # ds_interp = ds * np.nan
-        # else:
-            # ds_interp = ds.interpolate_na(
-                # dim="time", method="quadratic", max_gap=MAX_GAP,
-                # fill_value="extrapolate"
-            # )
-        # self.ds_interp = ds_interp
-        # self.fs = get_sampling_frequency(self.ds)
-
-    # def cross_spectral_matrix(self) -> xr.Dataset:
-
-        # # extract variables from dataset
-        # x, y, z = (
-            # ds_interp["eastward_displacement"],
-            # ds_interp["northward_displacement"],
-            # ds_interp["surface_elevation"]
-        # )
-
-        # # constants
-        # detrend = "constant"
-        # nperseg = min(NPERSEG, len(ds.time))
-
-        # # auto-spectra
-        # frq, Sxx = signal.welch(x, fs=self.fs, nperseg=nperseg, detrend=detrend)
-        # frq, Syy = signal.welch(y, fs=self.fs, nperseg=nperseg, detrend=detrend)
-        # frq, Szz = signal.welch(z, fs=self.fs, nperseg=nperseg, detrend=detrend)
-
-        # # cross-spectra
-        # frq, Sxz = signal.csd(x, z, fs=self.fs, nperseg=nperseg, detrend=detrend)
-        # frq, Syz = signal.csd(y, z, fs=self.fs, nperseg=nperseg, detrend=detrend)
-        # frq, Sxy = signal.csd(x, y, fs=self.fs, nperseg=nperseg, detrend=detrend)
-
-        # # find relative frequency
-        # fp =  np.trapz(frq * Szz**4, x=frq) / np.trapz(Szz**4, x=frq)
-        # ffp = frq / fp
-
-        # return  xr.Dataset(
-            # coords = {
-                # "frqs": ("frqs", frq),
-                # "ffp": ("frqs", ffp)
-            # },
-            # data_vars = {
-                # "Sxx": ("frqs", Sxx),
-                # "Syy": ("frqs", Syy),
-                # "Szz": ("frqs", Szz),
-                # "Sxz": ("frqs", Sxz),
-                # "Syz": ("frqs", Syz),
-                # "Sxy": ("frqs", Sxy),
-            # }
-        # )
-
-    # def directional_moments(self, Phi: xr.Dataset) -> xr.Dataset:
-        # Exx, Eyy, Ezz, Cxy, Qxz, Qyz = (
-            # np.real(Phi["Sxx"]), np.real(Phi["Syy"]), np.real(Phi["Szz"]),
-            # np.real(Phi["Sxy"]), np.imag(Phi["Sxz"]), np.imag(Phi["Syz"])
-        # )
-
-        # return  xr.Dataset(
-            # {
-                # "a1": Qxz / np.sqrt(Ezz * (Exx + Eyy)),
[…]
