### 2025-03-31 dspelaez/extended-wdm#6 @dspelaez
Hi @rcaneill. I just removed this file. I thought about adding more features, at some point, but it turned out to be unnecessary.
