### 2025-04-05 0b0e9c8
Paper was modified to include detailed information
