+++ joss/paper.bib
@@ -56,7 +56,7 @@
 }
 
 @techreport{Barstow_etal_2005,
-  type = {{COST} Office},
+  type = {COST Office},
   title = {Measuring and {{Analysing}} the Directional Spectrum of Ocean Waves},
   author = {Barstow, Stephen F and Bidlot, Jean-Raymond and Caires, Sofia and Donelan, Mark A and Drennan, William M and Dupuis, H{\'e}l{\`e}ne and Graber, Hans C and Green, J Jim and Gronlie, Oistein and Gu{\'e}rin, Christine and Gurgel, Klaus-Werner and G{\"u}nther, Heinz and Hauser, Dani{\`e}le and Hayes, Kenneth and Hessner, Katrin and Hoja, Danielle and Icard, Delphine and Kahma, Kimmo K and Keller, William C and Krogstad, Harald E and Lefevre, Jean-Michel and Lehner, Susanne and Magnusson, Anne Karin and Monbaliu, Jaak and Borge, Jose Carlos Nieto and Pettersson, Heidi and Plant, William J and Quentin, C{\'e}line Gwena{\"e}lle and Reichert, Konstanze and Reistad, Magnar and Rosenthal, Wolfgang and Saetra, Oyvind and {Schulz-Stellenfleth}, Johannes and Walsh, Edward J and Weill, Alain and Wolf, Judith and Wright, C Wayne and Wyatt, Lucy R},
   year = {2005},
