### 2025-04-07 #7942 @dspelaez
Hi @rcaneill, thanks for the heads up, I was indeed unsure how to proceed. Regarding the individual issues, should I close each one as I address it, or is it you who closes them?
