### 2025-04-07 76708e4
Quick example plot corrected
