### 2025-04-07 c070b8d
Quick example plot corrected. Changed size
