+++ CONTRIBUTING.md
@@ -1,0 +1,27 @@
+# Contributing to EWDM (Extended Wavelet Directional Method)
+
+Thank you for considering contributing to EWDM! We appreciate your interest in improving the package. Please take a moment to review the following guidelines before submitting your contributions.
+
+## How to Contribute
+1. Fork the repository.
+2. Create a new branch for your feature or bug fix.
+3. Make your changes.
+4. Ensure your code follows the project's coding style and conventions.
+5. Write tests to cover your changes.
+6. Update documentation as needed.
+7. Submit a pull request to the `main` branch of the original repository.
+
+## Code Style
+- Follow PEP 8 guidelines for Python code.
+- Use clear and descriptive variable and function names.
+- Comment your code to explain complex parts or algorithms.
+
+## Reporting Issues
+If you encounter bugs, issues, or have suggestions for improvements, please open an issue o GitHub. Provide a clear description of the problem, steps to reproduce, expected behavior, and any relevant information.
+
+## Contact
+The [discussion page](https://github.com/dspelaez/extended-wdm/discussion) can be used to ask questions and submit ideas.
+
+The best place to report a bug or to propose any enhancements is the [issue tracker](https://github.com/dspelaez/extended-wdm/issues) of EWDM's Github repository.
+
+Thank you!
