### 2025-04-07 #7942 @rcaneill
> Hi [@rcaneill](https://github.com/rcaneill), thanks for the heads up, I was indeed unsure how to proceed. Regarding the individual issues, should I close each one as I address it, or is it you who closes them?

You can close them, I’ll check in any case ;)
