### 2025-04-07 d200f02
JOSS badged added to README
