+++ joss/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 # Summary
 
-The accurate estimation of directional wave spectra is crucial for understanding ocean wave dynamics, air-sea interactions, and coastal processes [@Barstow_etal_2005]. Traditional methods, based on conventional Fourier transform, such as the Truncated Fourier Series (TFS), Maximum Likelihood (MLM) or Maximum Entropy (MEM) methods, often suffer from limitations in spectral resolution, inaccurate estimations and spurious spectral peaks. This is generally attributed to certain simplifications on the shape of the directional distribution and assumptions such as stationarity, which may not accurately capture the complexity of ocean waves [@Benoit_etal_1997; @PelaezZapata_2024a; @Barstow_etal_2005]. Wavelet-based methods provide a more flexible approach, thanks to its time-frequency decomposition capabilities [@Donelan_etal_1996; @PelaezZapata_2024a; @Krogstad_etal_2006]. This paper presents the EWDM (Extended Wavelet Directional Method), a Python toolkit developed to estimate the directional spectra of ocean waves using the Continuos Wavelet Transform (CWT). EWDM aims to address the limitations of conventional techniques providing a robust estimation of the directional wave spectrum from both single-point triplet data [@PelaezZapata_2024a; @Krogstad_etal_2006] and spatially-distributed arrays, extending the capabilities of the original WDM [@Donelan_etal_1996], which was initially proposed to work only on arrays. Consequently, EWDM can be used on diverse sources of data, including GPS buoys, pitch-roll buoys, arrays of wave staffs, stereo-images and acoustic Doppler current profilers (ADCP). Key features of the EWDM include the implementation of wavelet-based algorithms for extracting directional information from wave time series, improved estimation of wave directional distribution using Kernel Density Estimation (KDE), tools for processing and visualizing directional wave data, and compatibility with popular data sources such as the so-called Spotter buoys and the directional waveriders (DWR) from the CDIP (Coastal Data Information Program) database. The package is powered by `xarray` [@Hoyer_2017] labelled multi-dimensional arrays, which enhances its efficiency and scalability.
[…]
