### 2025-04-08 #7942 @rcaneill
> Hi [@rcaneill](https://github.com/rcaneill). I think I've addressed all the comments. Please take a look whenever you get a chance and let me know if any further improvements or changes are needed. Thanks a million.

Hi @dspelaez, almost all good for me. The code is ok, about the paper, I appreciate the new introduction! I opened a PR with minor language changes.
When it is corrected, my review will be done :)
