### 2025-04-09 dspelaez/extended-wdm#9 @miniufo
[JOSS review] add a demo
Hi @dspelaez

This is an issue for JOSS review at https://github.com/openjournals/joss-reviews/issues/7942

The EWDM provide a nice Quick example on how to use the package.  But I am curious if the authors could provide an example on showing the advantages of their wavelet method over traditional (fft or fitting) method.  In the paper, the wavelet method adopted is claimed to be superior to FFT or fitting under some conditions.  It would be readlly great if there is such an example demonstrating these kinds of situations.

I find in the Gallery [a comparison between the wavelet and fft methods](https://extended-wdm.readthedocs.io/en/latest/auto_gallery/plot_fourier_vs_wavelet.html).  But there is no explicit explanation on which one is better.  So I would suggest the authors to add an example demonstrating the advantages explicitly, and put a link in the README.  This should be helpful for attracting potential users.
