### 2025-04-09 dspelaez/extended-wdm#10 @miniufo
[JOSS review] add a section on state of the field
Hi @dspelaez

This is an issue for JOSS review at https://github.com/openjournals/joss-reviews/issues/7942

There is an iterm on my checklist:
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

In the paper, there is no clear review of the current state of the field.  After a quick search I find a similar package [wavespectra](https://github.com/wavespectra/wavespectra), which may do a similar job.  So I would suggest the authors to at least go through some of these the widely-used or well-documented packages, and explain the differences of EWDM from these packages in both the paper and the README.

This also help the potential users in selecting their tools.
