+++ joss/paper.md
@@ -36,7 +36,7 @@ The EWDM (Extended Wavelet Directional Method) is a Python toolkit developed to 
 
 # Statement of need
 
-Various aspects of physical oceanography, including air-sea interactions, wave-induced mixing and upper layer dynamics, as well as several applications in coastal engineering, such as, wave forecasting, prediction of rogue waves, quantification of coastal erosion, and the design of offshore structures and renewable energy devices, require accurate and precise knowledge of the directional distribution of ocean waves [@Barstow_etal_2005]. The directional wave spectrum provide essential information about the distribution of wave energy with respect to frequency and direction. However, the dynamic and unpredictable nature of ocean waves, coupled with the constraints of current measurement technologies and analysis methods, pose challenges in attaining accurate directional information.
+Various aspects of physical oceanography, including air-sea interactions, wave-induced mixing and upper layer dynamics, as well as several applications in coastal engineering, such as wave forecasting, prediction of rogue waves, quantification of coastal erosion, and the design of offshore structures and renewable energy devices, require accurate and precise knowledge of the directional distribution of ocean waves [@Barstow_etal_2005]. The directional wave spectrum provide essential information about the distribution of wave energy with respect to frequency and direction. However, the dynamic and unpredictable nature of ocean waves, coupled with the constraints of current measurement technologies and analysis methods, pose challenges in attaining accurate directional information.
 
 The directional wave spectra can be estimated from measurements using either spatial arrays of wave gauges or single-point triplets, where three perpendicular wave quantities are measured at the same point [@Barstow_etal_2005]. This latter is for instance the case of wave buoys, which are the most wide-spread means for obtaining _in-situ_ wave data. However, current methods rely either on truncated Fourier series -that are unable to capture complex sea states- or on statistical fitting of the distribution of wave directions and often involve assumptions that may not hold in real-world conditions. See @PelaezZapata_2024a, @Benoit_1993, @Benoit_1993a, @Benoit_and_Teisson_1995 and @Benoit_etal_1997 for a comprehensive review of various methods.
 
