### 2025-04-11 #7942 @rwegener2
Thanks to both reviewers for their thoughtful contributions!

Since I see some discussion of contributing guidelines I wanted to drop in with a few notes I have:
* One thing that is typically included in contributing guidelines is steps for someone to setup the development environment in case they want to contribute code to the open source project. This usually involves setting up an environment with an editable installation (ex. `pip install -e .`), installing developer only packages like `pytest`, and steps for running the tests. The readme is a good place to put an overview, but it's also really common to have "Contributor's Guide" as a page in the documentation. Contributor guidelines can get pretty elaborate (ex. [scipy](https://docs.scipy.org/doc/scipy/dev/contributor/contributor_toc.html)), but in this case I suggest including in a page of the documentation (or the readme) at least 1) a description for how to setup the developer environment and 2) steps to run the tests.
* a small thing: the docs link in the [Futher Examples](https://github.com/dspelaez/extended-wdm?tab=readme-ov-file#further-examples) section of the readme returns a 404

Let me know if you have any questions!
