+++ joss/paper.bib
@@ -101,6 +101,21 @@
   urldate = {2023-07-04}
 }
 
+@article{Ochoa_and_Delgado-Gonzalez_1990,
+  title = {Pitfalls in the Estimation of Wind Wave Directional Spectra by Variational Principles},
+  author = {Ochoa, Jos{\'e} and {Delgado-Gonz{\'a}lez}, Oscar E.},
+  year = {1990},
+  month = oct,
+  journal = {Applied Ocean Research},
+  volume = {12},
+  number = {4},
+  pages = {180--187},
+  issn = {01411187},
+  doi = {10.1016/S0141-1187(05)80025-2},
+  urldate = {2025-01-14},
+  copyright = {https://www.elsevier.com/tdm/userlicense/1.0/}
+}
+
 @article{Hoyer_2017,
   title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
   author    = {Hoyer, S. and J. Hamman},
@@ -111,4 +126,31 @@
   publisher = {Ubiquity Press},
   doi       = {10.5334/jors.148},
   url       = {https://doi.org/10.5334/jors.148}
+}
+
+@article{Johnson_2002_diwasp,
+  title={DIWASP, a directional wave spectra toolbox for MATLAB{\textregistered}: User Manual},
+  author={Johnson, David},
+  journal={Center for Water Research, University of Western Australia},
+  pages={1--23},
+  year={2002}
+}
+
+@inproceedings{Brodtkorb_2000_wafo,
+  title={WAFO-a Matlab toolbox for analysis of random waves and loads},
+  author={Brodtkorb, P{\"a}r Andreas and Johannesson, Par and Lindgren, Georg and Rychlik, Igor and Ryd{\'e}n, Jesper and Sj{\"o}, Eva},
+  booktitle={ISOPE International Ocean and Polar Engineering Conference},
+  pages={ISOPE--I},
+  year={2000},
+  organization={isope}
+}
+
+@article{Hafner_2021_fowd,
+  title={FOWD: A free ocean wave dataset for data mining and machine learning},
+  author={H{\"a}fner, Dion and Gemmrich, Johannes and Jochum, Markus},
+  journal={Journal of Atmospheric and Oceanic Technology},
+  volume={38},
+  number={7},
+  pages={1305--1322},
+  year={2021}
 }
