+++ README.md
@@ -9,23 +9,12 @@
 
 # EWDM: Extended Wavelet Directional Method
 
-Welcome to the GitHub repository for the **Extended Wavelet Directional Method
-(EWDM)** - a toolkit for estimating directional spectra of ocean waves using the
-Continuous Wavelet Transform (CWT).
+Welcome to the GitHub repository for the **Extended Wavelet Directional Method (EWDM)** - a toolkit for estimating directional spectra of ocean waves using the Continuous Wavelet Transform (CWT).
 
-This algorithm is an extension of the original WDM proposed by [Donelan et al.
-(1985)](10.1175/1520-0485(1996)026<1901:naotdp>2.0.co;2) that was only suitable
-for spatially-distributed arrays of wave staffs. This extension allows the
-estimation of the directional wave spectrum for GPS wave buoys, pitch-roll-heave
-buoys and Acoustic Doppler Current Profilers.
+This algorithm is an extension of the original WDM proposed by [Donelan et al. (1985)](10.1175/1520-0485(1996)026<1901:naotdp>2.0.co;2) that was only suitable for spatially-distributed arrays of wave staffs. This extension allows the estimation of the directional wave spectrum for GPS wave buoys, pitch-roll-heave buoys and Acoustic Doppler Current Profilers.
 
-The wavelet-based methods have emerged as a practical alternative to the
-conventional Fourier-based techniques, particularly well-suited for the analysis
-of data from [Spotter buoys](https://www.sofarocean.com/products/spotter) as
-demonstrated by [Peláez-Zapata et al.
-(2024)](https://doi.org/10.1175/JTECH-D-23-0058.1) and from other triplet data
-such as wave slopes as shown by [Krogstad et al.
-(2006)](https://onepetro.org/IJOPE/article-abstract/28936/Wavelet-And-Local-Directional-Analysis-of-Ocean?redirectedFrom=fulltext).
+The wavelet-based methods have emerged as a practical alternative to the conventional Fourier-based techniques, particularly well-suited for the analysis of data from [Spotter buoys](https://www.sofarocean.com/products/spotter) as demonstrated by [Peláez-Zapata et al. (2024)](https://doi.org/10.1175/JTECH-D-23-0058.1) and from other triplet data such as wave slopes as shown by [Krogstad et al. (2006)](https://onepetro.org/IJOPE/article-abstract/28936/Wavelet-And-Local-Directional-Analysis-of-Ocean?redirectedFrom=fulltext).
+
 
 **Key features** of the **EWDM** include:
 
@@ -33,26 +22,18 @@ such as wave slopes as shown by [Krogstad et al.
   information from wave time series.
 * Tools for processing and visualising directional wave data.
 * Powered by `xarray` labelled multi-dimensional arrays.
-* Helper functions to handle commonly used data sources such as 
+* Helper functions to handle commonly used data sources such as
   [Spotter buoys](https://www.sofarocean.com/products/spotter) and
   [CDIP](https://cdip.ucsd.edu/).
 * Documentation, examples, and comparison with conventional methods.
 
 
-Whether you are a researcher, student, or engineer in physical oceanography, 
-**EWDM** provides a powerful, user-friendly toolkit
-for in-depth analysis of directional ocean wave spectra. Join us in exploring
-the fascinating world of directional wave analysis and making meaningful
-contributions to the understanding of ocean wave dynamics. We welcome
-contributions, feedback, and collaboration from the community to further enhance
-the capabilities of **EWDM**.
-
[…]
