### 2025-04-15 1bcb405
Few citations fixed for joss paper. Statement of need was reworwed.
