+++ joss/paper.bib
@@ -1,4 +1,4 @@
-@article{PelaezZapata_2024a,
+@article{PelaezZapata2024a,
   title = {Ocean {{Wave Directional Distribution}} from {{GPS Buoy Observations}} off the {{West Coast}} of {{Ireland}}: {{Assessment}} of a {{Wavelet-Based Method}}},
   shorttitle = {Ocean {{Wave Directional Distribution}} from {{GPS Buoy Observations}} off the {{West Coast}} of {{Ireland}}},
   author = {{Pel{\'a}ez-Zapata}, Daniel and Pakrashi, Vikram and Dias, Fr{\'e}d{\'e}ric},
@@ -10,7 +10,7 @@
 }
 
 
-@inproceedings{Benoit_1993,
+@inproceedings{Benoit1993,
   title = {Extensive Comparison of Directional Wave Analysis Methods from Gauge Array Data},
   booktitle = {Proc. {{Intern}}. {{Symp}}. on {{Ocean Wave Measurement}} and {{Analysis}}},
   author = {Benoit, Michel},
@@ -18,7 +18,7 @@
   pages = {740--754}
 }
 
-@inproceedings{Benoit_1993a,
+@inproceedings{Benoit1993a,
   title = {Practical {{Comparative Performance Survey}} of {{Methods Used}} for {{Estimating Directional Wave Spectra}} from {{Heave-Pitch-Roll Data}}},
   booktitle = {Coastal {{Engineering}} 1992},
   author = {Benoit, Michel},
@@ -32,7 +32,7 @@
   isbn = {978-0-87262-933-2 978-0-7844-7953-7}
 }
 
-@inproceedings{Benoit_and_Teisson_1995,
+@inproceedings{Benoit1995,
   title = {Laboratory {{Comparison}} of {{Directional Wave Measurement Systems}} and {{Analysis Techniques}}},
   booktitle = {Coastal {{Engineering}} 1994},
   author = {Benoit, Michel and Teisson, Charles},
@@ -46,7 +46,7 @@
   isbn = {978-0-7844-0089-0 978-0-7844-7952-0}
 }
 
-@inproceedings{Benoit_etal_1997,
+@inproceedings{Benoit1997,
   title = {Analyzing Multidirectional Wave Spectra: A Tentative Classification of Available Methods},
   booktitle = {Proceedings of the 27th {{IAHR Congress}}},
   author = {Benoit, Michel and Frigaard, Peter and Sh{\"a}ffer, Hemming A.},
@@ -55,7 +55,7 @@
   address = {San Francisco, CA, USA}
 }
 
-@techreport{Barstow_etal_2005,
+@techreport{Barstow2005,
   type = {COST Office},
   title = {Measuring and {{Analysing}} the Directional Spectrum of Ocean Waves},
[…]
