+++ joss/paper.bib
@@ -90,7 +90,7 @@
 
 @article{Donelan1996,
   title = {Nonstationary Analysis of the Directional Properties of Propagating Waves},
-  author = {Donelan, M. A. and Drennan, W. M. and Magnusson, A. K.},
+  author = {Donelan, M. and Drennan, W. M. and Magnusson, A. K.},
   year = {1996},
   month = sep,
   journal = {Journal of Physical Oceanography},
