### 2025-04-15 a0ea545
Minor fixes on joss.md paper
