### 2025-04-15 758bcfd
Contributing guide added to the documentation
