+++ docs/contribute.rst
@@ -1,0 +1,4 @@
+How to contribute?
+==================
+
+Thank you for considering contributing to EWDM! We appreciate your interest in improving the package. Please take a moment to review the following guidelines before submitting your contributions.
