### 2025-04-16 15599b7
Some adjustements in th README.md
