### 2025-04-16 de85cae
Adding new example comparing different methods
