### 2025-04-16 a968023
Adding new example comparing different methods
