+++ docs/gallery/plot_comparison_of_methods.py
@@ -1,0 +1,253 @@
+#! /usr/bin/env python
+# -*- coding: utf-8 -*-
+# vim:fenc=utf-8
+
+"""
+Comparison of different methods
+===============================
+
+"""
+
+import numpy as np
+import xarray as xr
+import scipy.signal as signal
+
+from matplotlib import pyplot as plt
+
+import ewdm
+from ewdm.sources import SpotterBuoysDataSource
+from ewdm.plots import plot_directional_spectrum
+
+import warnings
+warnings.filterwarnings("ignore", category=RuntimeWarning)
+
+# %% 
+# Load sample dataset
+# ---------------------------------------
+#
+spotter = SpotterBuoysDataSource("../../data/displacement.csv")
+dataset = (
+    spotter
+    .read_dataset()
+    .sel(time=slice("2020-08-20 10:00", "2020-08-20 10:30"))
+)
+print(dataset)
+
+# plot time series of the first 5 minutes
+fig, (ax1, ax2, ax3) = plt.subplots(
+    3,1, figsize=(6,6), sharex=True, sharey=True
+)
+subset = dataset.sel(time=slice("2020-08-20T10:00", "2020-08-20T10:05"))
+subset["surface_elevation"].plot(ax=ax1)
+subset["eastward_displacement"].plot(ax=ax2)
+subset["northward_displacement"].plot(ax=ax3)
+_ = ax1.set(xlabel="", ylabel="$\\eta(t)$ [m]")
+_ = ax2.set(xlabel="", ylabel="$x(t)$ [m]")
+_ = ax3.set(xlabel="", ylabel="$y(t)$ [m]")
+
+
+# %% 
+# Classic directional spectral analysis
+# -------------------------------------
+#
+class ClassicSpectralAnalysis(object):
+    """This class implements the classic directional spectral analysis"""
+
+    def __init__(self, dataset: xr.Dataset, fs: float=1.0, nperseg=256):
+        self.dataset = dataset # time series
+        self.fs = fs # sampling frequency
+        self.nperseg = nperseg # number of fourier components
+
+    def cross_spectral_matrix(self) -> xr.Dataset:
+
+        # extract variables from dataset
+        x, y, z = (
+            self.dataset["eastward_displacement"],
+            self.dataset["northward_displacement"],
+            self.dataset["surface_elevation"]
+        )
+
+        # constants
+        fft_args = {
+            "fs": self.fs,
+            "detrend": "constant",
+            "nperseg": min(self.nperseg, len(self.dataset["time"])),
+        }
+
+        # auto-spectra
+        frq, Sxx = signal.welch(x, **fft_args)
+        frq, Syy = signal.welch(y, **fft_args)
+        frq, Szz = signal.welch(z, **fft_args)
+
+        # cross-spectra
+        frq, Sxz = signal.csd(x, z, **fft_args)
+        frq, Syz = signal.csd(y, z, **fft_args)
+        frq, Sxy = signal.csd(x, y, **fft_args)
+
+        return  xr.Dataset(
+            coords = {
+                "frequency": ("frequency", frq),
+            },
+            data_vars = {
+                "Sxx": ("frequency", Sxx),
+                "Syy": ("frequency", Syy),
+                "Szz": ("frequency", Szz),
+                "Sxz": ("frequency", Sxz),
+                "Syz": ("frequency", Syz),
[…]
