### 2025-04-17 a067087
Working on detailed example on methods comparison
