+++ docs/gallery/plot_comparison_of_methods.py
@@ -6,6 +6,24 @@
 Comparison of different methods
 ===============================
 
+It is a common practice to express the directional wave spectrum,
+:math:`E(f,\\theta)`, as the product of two functions:
+
+    .. math:: E(f,\\theta) = S(f) D(f,\\theta)
+
+where :math:`S(f)` is the frequency spectrum and :math:`D(f,\\theta)` is the
+directional distribution function. The problem basically consist of
+finding :math:`D(f,\\theta)`.
+
+This section compares different methods for estimating the directional
+distribution :math:`D(f,\\theta)` and the the directional wave
+spectra :math:`E(f,\\theta)`. We compare the wavelet-based directional method
+implemented in EWDM with two of the conventional methods based on Fourier
+spectral analysis, namely Truncated Fourier Series (TSF) and Maximum Entropy
+Method (MEM).
+
+The results are visualised and compared to understand the performance and
+differences between these methods.
 """
 
 import numpy as np
@@ -21,9 +39,15 @@ from ewdm.plots import plot_directional_spectrum
 import warnings
 warnings.filterwarnings("ignore", category=RuntimeWarning)
 
-# %% 
+# %%
 # Load sample dataset
 # ---------------------------------------
+# For this example, were use data from a Spotter buoy deployed off
+# the west coast of Ireland in August 2020. In particular, we are using
+# a sample of 30 minutes.
+#
+# The dataset contains the two horizontal wave-induced displacements and
+# the vertical sea surface elevation.
 #
 spotter = SpotterBuoysDataSource("../../data/displacement.csv")
 dataset = (
@@ -32,6 +56,12 @@ dataset = (
     .sel(time=slice("2020-08-20 10:00", "2020-08-20 10:30"))
 )
 print(dataset)
+
+# %%
+# Let's plot a subsample of 5 minutes of these time series to have
+# an idea on what they look like. As it can be seen, the amplitude of
+# the wave-induced buoy movement, in both the horizontal and vertical
+# coordinates, is about 3-4 metres.
 
 # plot time series of the first 5 minutes
 fig, (ax1, ax2, ax3) = plt.subplots(
@@ -46,10 +76,52 @@ _ = ax2.set(xlabel="", ylabel="$x(t)$ [m]")
 _ = ax3.set(xlabel="", ylabel="$y(t)$ [m]")
 
 
-# %% 
+# %%
 # Classic directional spectral analysis
 # -------------------------------------
+# The following class implements the conventional cross-spectral analysis
+# using Fourier transform. This class takes in a dataset containing eastward
+# and northward displacements along with the surface elevation data. The
+# input parameters are sampling frequency of the time series and number of
+# Fourier components for analysis.
 #
+# Considering the three time series, :math:`x(t)`, :math:`y(t)` and
+# :math:`z(t)` (We normally use :math:`\eta(t)` for the vertical but it is
+# changed here to :math:`z(t)` for simplicity and convenience), the
+# cross-spectral matrix is computed as:
+#
+# .. math:: \Phi(f) = \begin{bmatrix}
+#                        S_{xx} & S_{xy} & S_{xz} \\
+#                        S_{yx} & S_{yy} & S_{yz} \\
[…]
