### 2025-04-17 60aa004
Adding a discussion on differences between methods
