### 2025-04-17 30f6725
README.md updated to mention similar projects
