### 2025-04-17 52c9a67
README.md updated to mention similar projects
