### 2025-04-17 09e0416
README.md updated to mention similar projects
