### 2025-04-17 dspelaez/extended-wdm#10 @dspelaez
Hi @miniufo

Thanks for this comment. I fully agree with you. I included a detailed description of the `State of the field` in the paper, mentioning some of these relevant projects and how they could be potentially compatible with EWDM. I also briefly mention this in the README.
