### 2025-04-17 dspelaez/extended-wdm#9 @dspelaez
This is a great suggestion. I included a [detailed example](https://extended-wdm.readthedocs.io/en/latest/auto_gallery/plot_comparison_of_methods.html) in the documentation, explaining the differences between some of the most commonly used methods, and comparing their outputs with EWDM. I also cited relevant papers in which similar comparisons have been done. I would like to mention here that determining 'the best method' is not an easy task, some of them work well under certain conditions, some others not that well, and so on. Our goal with EWDM is implementing this alternative wavelet-based method, so users can have a wider range of options to choose.
