### 2025-04-17 #7942 @dspelaez
Hi @miniufo, @rcaneill and @rwegener2

Thanks a million for your great comments and suggestions, they really helped improve the overall quality and value of this software. I've gone through your feedback, so feel free to take a look when you guys get a chance.
