### 2025-04-17 9cf0c7c
CITATION.cff file added
