### 2025-04-17 014c119
Small adjustements in comparison example
