### 2025-04-17 b65cb33
Small adjustements in comparison example
