### 2025-04-18 #7942 @miniufo
That's great.  I just go through all the modifications.  They looks good to me now.  So what's next?  Maybe we need to re-generate the rendered PDF file for the paper?
