### 2025-04-18 ea01c27
Fixing minor typos in the documentation
