### 2025-04-19 #7942 @rcaneill
https://github.com/dspelaez/extended-wdm/commit/bceb8b62d63a2563bfb21d7b24be90aecad8974f

This commit deleted part of the intro, @dspelaez can you fix this?
