+++ joss/paper.md
@@ -30,8 +30,7 @@ bibliography: paper.bib
 
 # Summary
 
-uted to certain simplifications on the shape of the directional distribution and assumptions such as stationarity, which may not always capture the complexity of ocean waves [@Benoit1997]. Wavelet-based methods provide a more flexible approach, thanks to their time-frequency decomposition capabilities [@Donelan1996; @Donelan2015; @PelaezZapata2024a; @Krogstad2006]. This paper presents the EWDM (Extended Wavelet Directional Method), a Python toolkit developed to estimate the directional spectrum of ocean waves based on a wavelet-based technique. EWDM aims to address the limitations of conventional techniques by providing a robust estimation of the directional wave spectrum from both single-point triplet data [@PelaezZapata2024a; @Krogstad2006] and spatially-distributed arrays [@Donelan1996; @Donelan2015]. Consequently, EWDM can be used on diverse sources of data, including GPS buoys, pitch-roll buoys, arrays of wave staffs, acoustic Doppler current profilers (ADCP) and sampled points from stereo-videos of the sea surface. Key features of the EWDM include the implementation of wavelet-based algorithms for extracting directional information from wave time series, improved estimation of wave directional distribution using Kernel Density Estimation (KDE), tools for processing and visualizing directional wave data, and compatibility with popular data sources, including Spotter buoys and CDIP (Coastal Data Information Program) database. The package is powered by [`xarray`](https://github.com/pydata/xarray)'s [@Hoyer2017] labelled multi-dimensional arrays, which enhances its efficiency, scalability and compatibility with scientific tools and workflows.
-
+The accurate estimation of directional wave spectra is crucial for understanding ocean wave dynamics, air-sea interactions, and coastal processes [@Barstow2005]. The directional wave spectrum is often computed by resolving the distribution of wave energy as a function of frequency and direction using mathematical methods applied to time series from measuring instruments (e.g., bouys). These time series typically consist of either triplets of variables at a single point (e.g., velocity or acceleration components) or arrays of measurements distributed across multiple spatial locations (e.g., wave staffs). Traditional methods, based on the conventional Fourier cross-spectral analysis, such as the Truncated Fourier Series (TFS), Maximum Likelihood (MLM), or Maximum Entropy (MEM) methods, often suffer from limitations in spectral resolution, inaccurate estimations, and spurious spectral peaks. This disadvantage is generally attributed to certain simplifications on the shape of the directional distribution and assumptions such as stationarity, which may not always capture the complexity of ocean waves [@Benoit1997]. Wavelet-based methods provide a more flexible approach, thanks to their time-frequency decomposition capabilities [@Donelan1996; @Donelan2015; @PelaezZapata2024a; @Krogstad2006]. This paper presents the EWDM (Extended Wavelet Directional Method), a Python toolkit developed to estimate the directional spectrum of ocean waves based on a wavelet-based technique. EWDM aims to address the limitations of conventional techniques by providing a robust estimation of the directional wave spectrum from both single-point triplet data [@PelaezZapata2024a; @Krogstad2006] and spatially-distributed arrays [@Donelan1996; @Donelan2015]. Consequently, EWDM can be used on diverse sources of data, including GPS buoys, pitch-roll buoys, arrays of wave staffs, acoustic Doppler current profilers (ADCP) and sampled points from stereo-videos of the sea surface. Key features of the EWDM include the implementation of wavelet-based algorithms for extracting directional information from wave time series, improved estimation of wave directional distribution using Kernel Density Estimation (KDE), tools for processing and visualizing directional wave data, and compatibility with popular data sources, including Spotter buoys and CDIP (Coastal Data Information Program) database. The package is powere
[…]
