+++ docs/gallery/plot_stereo_video.py
@@ -5,7 +5,6 @@
 """
 Estimating directional spectra from stereo-imaging
 ==================================================
-
 Stereo-imaging has emerged as a unique measurement technique for obtaining
 precise directional information about wave fields. One of its key advantages
 lies in the direct estimation of directional wave spectra through
@@ -97,10 +96,10 @@ nc_obj = nc.Dataset(LOCAL_VIDEO_FILE)
 # To pick the pixels, we define the corresponding indices. We are going to
 # evaluate four different configurations. 
 #
-# - The optimal array proposed by Young et al (1994).
-# - A simple pentagon.
-# - 10 random points.
-# - 20 random points.
+# 1. The optimal array proposed by Young et al (1994).
+# 2. A simple pentagon.
+# 3. 10 random points.
+# 4. 20 random points.
 
 indices1 = [
     (95,95), (95,175), (95,15), (175,15), (15,15),
