### 2025-04-21 #7942 @rwegener2
Thanks for the updates @miniufo and @rcaneill. Would you both mind checking off any remaining on items on your checklist to confirm that you see these items as fulfilled by this package?

Also, thanks for the changes to the Contributing Guide, @dspelaez, it is thorough and looks beautiful!
