### 2025-04-22 7ab75e2
Added DOI field for a reference in the paper.bib file.
