+++ joss/paper.bib
@@ -166,6 +166,7 @@
   journal={Journal of Atmospheric and Oceanic Technology},
   volume={38},
   number={7},
+  doi={10.1175/jtech-d-20-0185.1},
   pages={1305--1322},
   year={2021}
 }
