### 2025-04-22 4e8b006
Minor typos and redundancies fixed in paper.md
