+++ joss/paper.md
@@ -45,7 +45,7 @@ Available wave sensors typically measure at single-point locations (e.g. buoys) 
 
 Numerous software tools are available for analysing ocean wave data, each designed for specific applications and needs. Well-established tools, such as [`WAFO`](https://www.maths.lth.se/matstat/wafo/) [Wave Analysis for Fatigue and Oceanography, @Brodtkorb2000wafo], focus on the statistical analysis of random waves, fatigue analysis, wave-induced loads, and reliability assessments in ocean engineering. Alternative, recent software packages, such as [`FOWD`](https://github.com/dionhaefner/FOWD) [Free Ocean Wave Dataset for Data Mining and Machine Learning, @Hafner2021fowd], provide methods for processing wave elevation time series, including the estimation of zero-crossing parameters, such as crest-to-trough height, skewness and kurtosis; and spectral parameters, such as significant wave height, peak and mean periods, directional spreading and spectral bandwidths. [`FOWD`](https://github.com/dionhaefner/FOWD) focuses on providing an extensive dataset of wave parameters optimised for machine learning and data mining applications. While these packages facilitate the estimation of certain directional parameters, they do not have the functionality to calculate the directional spectrum from wave observations.
 
-[`DIWASP`](https://github.com/metocean/diwasp) [DIrectional WAve SPectrum analysis, @Johnson2002diwasp], is a widely adopted MATLAB-based toolbox specifically designed for this task. [`DIWASP`](https://github.com/metocean/diwasp) implements a variety of routines for directional wave spectrum estimation, such as the Iterative Maximum Likelihood Method (IMLM), the Extended Maximum Entropy Principle (EMEP), the Bayesian Directional Method (BDM), among others. However, all these methods rely on conventional Fourier cross-spectral analysis, and therefore, are subject to the limitations previously discussed.
+[`DIWASP`](https://github.com/metocean/diwasp) [DIrectional WAve SPectrum analysis, @Johnson2002diwasp] is a widely adopted MATLAB-based toolbox specifically designed for this task. [`DIWASP`](https://github.com/metocean/diwasp) implements a variety of routines for directional wave spectrum estimation, such as the Iterative Maximum Likelihood Method (IMLM), the Extended Maximum Entropy Principle (EMEP), the Bayesian Directional Method (BDM), among others. However, all these methods rely on conventional Fourier cross-spectral analysis, and therefore, are subject to the limitations previously discussed.
 
 Modern Python-based tools, such as [`wavespectra`](https://github.com/wavespectra/wavespectra) and [`oceanwaves-python`](https://github.com/openearth/oceanwaves-python), provide an abstraction layer for typical directional wave spectra manipulation. These packages provide functionalities for common operations, such as reading and writing spectral data from numerical models (e.g., SWAN, WAVEWATCH III) and wave instruments (e.g., Spotter buoys, TRIAXYS, Datawell), spectral partitioning, estimation of spectral parameters, standard conversions and advanced plotting. They do not incorporate, however, algorithms for estimating the directional wave spectrum from raw wave measurements. EWDM complements the ecosystem, providing routines for estimating the directional wave spectrum directly from measurements of wave time series. The outputs obtained from EWDM can potentially be integrated with these tools to broaden its functionalities, opening up new possibilities to the scientific community.
 
