### 2025-04-23 99d1f43
Fixing some issues with references in paper.bib
