+++ joss/paper.bib
@@ -1,12 +1,16 @@
 @article{PelaezZapata2024a,
-  title = {Ocean {{Wave Directional Distribution}} from {{GPS Buoy Observations}} off the {{West Coast}} of {{Ireland}}: {{Assessment}} of a {{Wavelet-Based Method}}},
-  shorttitle = {Ocean {{Wave Directional Distribution}} from {{GPS Buoy Observations}} off the {{West Coast}} of {{Ireland}}},
-  author = {{Pel{\'a}ez-Zapata}, Daniel and Pakrashi, Vikram and Dias, Fr{\'e}d{\'e}ric},
+  title = {Ocean Wave Directional Distribution from GPS Buoy Observations off the West Coast of Ireland: Assessment of a Wavelet-Based Method},
+  volume = {41},
+  ISSN = {1520-0426},
+  url = {http://dx.doi.org/10.1175/JTECH-D-23-0058.1},
+  DOI = {10.1175/jtech-d-23-0058.1},
+  number = {8},
+  journal = {Journal of Atmospheric and Oceanic Technology},
+  publisher = {American Meteorological Society},
+  author = {Peláez-Zapata,  Daniel and Pakrashi,  Vikram and Dias,  Frédéric},
   year = {2024},
   month = aug,
-  doi = {10.1175/JTECH-D-23-0058.1},
-  urldate = {2024-09-16},
-  chapter = {Journal of Atmospheric and Oceanic Technology}
+  pages = {749–765}
 }
 
 
