### 2025-04-24 e7957d1
Binder description added to contributing guidelines
