### 2025-04-29 #7942 @rwegener2
Okay @dspelaez, I have completed my first set of checks. Please complete the tasks in section 2 of [this comment](https://github.com/openjournals/joss-reviews/issues/7942#issuecomment-2821360111).
