### 2025-04-29 #7942 @dspelaez
Hi @rwegener2. Thanks a million.

The new release is the version [v1.1](https://github.com/dspelaez/extended-wdm/releases/tag/v1.1)

It was archived in Zenodo and the DOI [10.5281/zenodo.15304371](https://doi.org/10.5281/zenodo.15304371) was assigned.

Please let me know if everything is ok.
