+++ .github/workflows/python-publish.yml
@@ -36,5 +36,5 @@ jobs:
     - name: Publish package
       uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
       with:
-        user: dspelaez
+        user: __token__
         password: ${{ secrets.PYPI_API_TOKEN }}
