### 2025-04-29 #7942 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
