### 2025-04-30 #7942 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "ireland"
