### 2025-04-30 9a64275
Fixing capitalization in paper.md
