+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @article{PelaezZapata2024a,
-  title = {Ocean Wave Directional Distribution from GPS Buoy Observations off the West Coast of Ireland: Assessment of a Wavelet-Based Method},
+  title = {{Ocean Wave Directional Distribution from GPS Buoy Observations off the West Coast of Ireland: Assessment of a Wavelet-Based Method}},
   volume = {41},
   ISSN = {1520-0426},
   url = {http://dx.doi.org/10.1175/JTECH-D-23-0058.1},
