+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ewdm"
-version = "1.0"
+version = "1.1"
 authors = [
   {name = "Daniel Peláez-Zapata", email="daniel.pelaez-zapata@ucdconnect.ie"},
 ]
