### 2025-04-30 ac3e9cf
Fixing capitalization in additional bibtex entries
