+++ joss/paper.bib
@@ -13,9 +13,8 @@
   pages = {749–765}
 }
 
-
 @inproceedings{Benoit1993,
-  title = {Extensive Comparison of Directional Wave Analysis Methods from Gauge Array Data},
+  title = {{Extensive Comparison of Directional Wave Analysis Methods from Gauge Array Data}},
   booktitle = {Proc. {{Intern}}. {{Symp}}. on {{Ocean Wave Measurement}} and {{Analysis}}},
   author = {Benoit, Michel},
   year = {1993},
@@ -51,7 +50,7 @@
 }
 
 @inproceedings{Benoit1997,
-  title = {Analyzing Multidirectional Wave Spectra: A Tentative Classification of Available Methods},
+  title = {{Analyzing Multidirectional Wave Spectra: A Tentative Classification of Available Methods}},
   booktitle = {Proceedings of the 27th {{IAHR Congress}}},
   author = {Benoit, Michel and Frigaard, Peter and Sh{\"a}ffer, Hemming A.},
   year = {1997},
@@ -61,14 +60,14 @@
 
 @techreport{Barstow2005,
   type = {COST Office},
-  title = {Measuring and {{Analysing}} the Directional Spectrum of Ocean Waves},
+  title = {{Measuring and Analysing the Directional Spectrum of Ocean Waves}},
   author = {Barstow, Stephen F and Bidlot, Jean-Raymond and Caires, Sofia and Donelan, Mark A and Drennan, William M and Dupuis, H{\'e}l{\`e}ne and Graber, Hans C and Green, J Jim and Gronlie, Oistein and Gu{\'e}rin, Christine and Gurgel, Klaus-Werner and G{\"u}nther, Heinz and Hauser, Dani{\`e}le and Hayes, Kenneth and Hessner, Katrin and Hoja, Danielle and Icard, Delphine and Kahma, Kimmo K and Keller, William C and Krogstad, Harald E and Lefevre, Jean-Michel and Lehner, Susanne and Magnusson, Anne Karin and Monbaliu, Jaak and Borge, Jose Carlos Nieto and Pettersson, Heidi and Plant, William J and Quentin, C{\'e}line Gwena{\"e}lle and Reichert, Konstanze and Reistad, Magnar and Rosenthal, Wolfgang and Saetra, Oyvind and {Schulz-Stellenfleth}, Johannes and Walsh, Edward J and Weill, Alain and Wolf, Judith and Wright, C Wayne and Wyatt, Lucy R},
   year = {2005},
   institution = {{European cooperation in the field of scientific and technical research}}
 }
 
 @article{Young1994,
-  title = {On the Measurement of Directional Wave Spectra},
+  title = {{On the Measurement of Directional Wave Spectra}},
   author = {Young, I.R.},
   year = {1994},
   month = jan,
@@ -83,7 +82,7 @@
 
 
 @inproceedings{Krogstad2006,
-  title = {Wavelet and Local Directional Analysis of Ocean Waves},
+  title = {{Wavelet and Local Directional Analysis of Ocean Waves}},
   booktitle = {The Sixteenth International Offshore and Polar Engineering Conference},
   author = {Krogstad, Harald Elias and Magnusson, Anne Karine and Donelan, Mark},
   year = {2006},
@@ -93,7 +92,7 @@
 
 
 @article{Donelan1996,
-  title = {Nonstationary Analysis of the Directional Properties of Propagating Waves},
+  title = {{Nonstationary Analysis of the Directional Properties of Propagating Waves}},
   author = {Donelan, M. and Drennan, W. M. and Magnusson, A. K.},
   year = {1996},
   month = sep,
@@ -106,7 +105,7 @@
 }
 
 @article{Donelan2015,
[…]
