### 2025-04-30 5e5fab4
Fixing capitalization in WAFO entry
