+++ joss/paper.bib
@@ -155,7 +155,7 @@
 }
 
 @inproceedings{Brodtkorb2000wafo,
-  title={WAFO-a Matlab toolbox for analysis of random waves and loads},
+  title={{WAFO - a MATLAB toolbox for analysis of random waves and loads}},
   author={Brodtkorb, P{\"a}r Andreas and Johannesson, Par and Lindgren, Georg and Rychlik, Igor and Ryd{\'e}n, Jesper and Sj{\"o}, Eva},
   booktitle={ISOPE International Ocean and Polar Engineering Conference},
   pages={ISOPE--I},
