### 2025-04-30 #7942 @dspelaez
Hi @kthyng

Thanks for spotting this. I fixed the capitalisation in all bibtex entries.
